refactor(core): make ProjectsPanel self-contained (PM lift, phase 1)
Decouple the entire PM modal/handler surface from CopperForgeApp so the
Projects panel can lift into copperforge-pro.
