+++ crates/copperforge-core/src/ui/mod.rs
@@ -4,6 +4,7 @@ pub mod drc_panel;
 pub mod grid_settings;
 pub mod view_settings_panel;
 pub mod projects_panel;
+pub mod projects_modals;
 pub mod settings_panel;
 pub mod about_panel;
 pub mod tabs;
@@ -15,6 +16,7 @@ pub use layer_controls::show_layers_panel;
 pub use drc_panel::show_drc_panel;
 pub use grid_settings::show_grid_panel;
 pub use projects_panel::show_projects_panel;
+pub use projects_modals::show_projects_modals;
 pub use settings_panel::show_settings_panel;
 pub use about_panel::AboutPanel;
 pub use bom_panel_v2::{show_bom_panel, BomPanelState};
